Refactor: 스케쥴아이템의 삭제 기능 분리
