+++ src/main/java/edu/example/wayfarer/service/ScheduleItemService.java
@@ -13,4 +13,5 @@ public interface ScheduleItemService {
 
     ScheduleItemResponseDTO update(ScheduleItemUpdateDTO scheduleItemUpdateDTO);
 
+    void delete(Long scheduleItemId);
 }
